backend/requirements.txt
# FastAPI and server
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Environment and configuration
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# LLM API
openai==1.57.2

# HTTP client
httpx==0.28.1

# CORS and middleware
python-multipart==0.0.20

# Firebase
firebase-admin==6.5.0
